+++ test/integration/039_config_test/models-get/any_model.sql
@@ -1,0 +1,2 @@
+-- models/any_model.sql
+select {{ config.get('made_up_nonexistent_key', 'default_value') }} as col_value